# Set channel.StandardSocketOptions.TCP_NODELAY to true