I was having trouble with latency ranging from 30ms to 150ms.  Setting TCP_NODELAY improved latency and it now ranges between 30 to 50ms.  I'm not sure if having this be true for everyone is a good idea.