@dan335 2017-04-24
I got the idea from issue https://github.com/TooTallNate/Java-WebSocket/issues/409