@marci4 2017-04-24
Hello @dan335,

thx for your pull request :)

This code introduces a dependency for Java 1.7.
Since we also support older android versions, I am also looking for a different approach.

Greetings
marci4